separate functions and autoload to simplify native fallback from composer autoload
